Moved Signal for PROG when using MEGA from Pin 4 to Pin 2
If compiled for MEGA, use OC3B instead of OC0B to generate DCC Signals
for Programming Track.  This changes output from Pin 4 to Pin 2.
Reason for change is to keep pin 4 clear so that MEGA can be used with
Ethernet Shield (that uses pin 4 for its SD Card).   No change if
compiled for UNO — prog signal is still produced with 0C0B on Pin 5.