Added controllable OUTPUT pins
Create/edit/delete/control OUTPUT pins for custom use.  Pins can be set
to either ACTIVE or INACTIVE state, where the default behavior is
ACTIVE=HIGH and INACTIVE=LOW.    This default behavior can reversed on
a pin-by-pin basis.  STATE of pin is retained in EEPROM and default
behavior is to restore upon power-up.  This behavior can be changed on
a pin-by-pin basis to always set STATE to ACTIVE or INACTIVE upon
power-up.