Clean-up #include for Ethernet Library
This update cleans up the way Ethernet Libraries are included.  It is
designed to be compiled under the latest Arduino IDE - version 1.6.6.
This latest version of the IDE correct a long-standing “bug” in which
#includes for libraries must be present in the main sketch file, even
if not used in that file.  With the new version, the Ethernet Library
is named in only one place: Config.h.  Also, the library is not even
included if you compile for the Serial Port only (and therefore the
name does not need to point to any real library).

NOTE:  This version will ONLY compile with the latest Arduino IDE 1.6.6
that contain the needed “fix.”