New version of CV reading and writing to Programming Track
The <R>, <W>, and <B> command that read and write CV values to
Programming Track have been re-written to (hopefully) be more robust.
Originally, current sensing simply average analogReads over a fixed
period of time and compared the average against a threshold.  This is
very sensitive to the timing to the decoder acknowledge and length of
the sample period.  In this new version we use an
exponentially-weighted average to look for a spike above an
pre-measures current baseline.   This should be insensitive the the
size of the sample window, the timing of the decoder acknowledgment,
and baseline current.