Created <D> diagnostic command
The <D> command increases the pre-scalers for Timer 0 and 1 (Uno) or
Timer 3 and 1 (Mega) as well as SLOWS DOWN THE SYSTEM CLOCK by a factor
of 256.  The purpose of this is enable a visual inspection of the DCC
signals by connecting to an eternal LED or the on-board LED attached to
pin 13.  Once a <D> command is issued, Serial Communication will be
disrupted.  This command is only to be used under the Serial protocol.
The board much be power-cycled or reset (either through the on-board
reset button to by closing and re-opening the Arduino Serial Monitor
Window) to return the clock rot normal.