1) Added code for LCD Throttle support.  Disabled in Config.h
2) Added support for Bonjour/Zeroconf registration.  Disabled in Config.h

WARNING: If using the EthernetBonjour library (which this does), you MUST
make the changes to the Ethernet library as outlined in this Arduino forum
thread: http://forum.arduino.cc/index.php?topic=234340.0

ALSO, you must change line 148 of EthernetBonjour.cpp to call
this->beginMulti() instead of this->beginMulticast() ... apparently
a typo still in the library.

FINALLY, you might want to change #define HAS_NAME_BROWSING in
EthernetBonjour.cpp to 0, to save some space, since we are registering
a service, not browsing for one.  I have not (yet) tested that though.